Generating dags
