+++ s2gos-airflow/dags/example_dag.py
@@ -1,4 +1,5 @@
 from airflow.sdk import DAG, task
+
 #from airflow.providers.standard.operators.python import PythonOperator
