+++ s2gos-server/src/s2gos_server/services/base/__init__.py
@@ -2,9 +2,11 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-from .service_base import DEFAULT_CONFORMANCES, ServiceBase
+from .function_process import FunctionProcess
+from .service_base import DEFAULT_CONFORMS_TO, ServiceBase
 
 __all__ = [
+    "FunctionProcess",
     "ServiceBase",
-    "DEFAULT_CONFORMANCES",
+    "DEFAULT_CONFORMS_TO",
 ]
